Create xss.py
We then capture the HTTP status code using the subprocess.check_output() method and check if it is equal to 200. If it is, then the payload was executed, and we have detected an XSS vulnerability. Otherwise, we assume that no XSS vulnerability was found.

Note that this script assumes that the webpage has at least one input field that accepts user input. It also assumes that the input field is of type application/x-www-form-urlencoded. If the webpage uses a different input format, you may need to modify the curl command accordingly.