Create XSS1.PY
This script uses the requests library to send a POST request to the specified URL with an injected payload. It then checks if the payload was reflected in the response, which indicates an XSS vulnerability. The script loops through a list of input fields to test for XSS. You can modify the url and input_fields variables to suit your needs.