Create broken-authentication.py
This script sends a POST request to the website's login page with a list of common usernames and passwords. It then checks if the login was successful by looking for the word "logout" in the response. If a valid set of credentials is found, it prints the credentials to the console. You can modify the base_url, usernames, and passwords variables to suit your needs.