Create SQLI.py
This script sends a GET request to the specified search URL with an injected search query that contains a SQL injection attack. It then checks if the SQL injection was successful by looking for common error messages in the response text such as 'error' or 'syntax error'. If the script finds one of these error messages in the response, it prints 'SQL injection vulnerability found'. Otherwise, it prints 'No SQL injection vulnerability found'.