Create ALL-OWASP-CHECK
This script tests for each vulnerability by performing a specific action on the website and checking the response for indicators of the vulnerability. You can modify the url variable to test for vulnerabilities on your own website. You can also add additional checks to test for other OWASP Top 10 vulnerabilities that are not included in this script.